remove previous implementation
As I was mulling this over this weekend I realized
why the JA-SIG implementation carries around
instances of these classes with their tickets
rather than setting the timeout as static and
using static methods on the Classes themselves.
The realization was that we would have to
duplicate a lot of logic in several classes for
the same type of expiration policy but with
different timeout lengths. But rather than
dragging instances of these classes around and
putting them in the DB we'll do something more
rubyish....

Signed-off-by: Tyler Pickett <t.pickett66@gmail.com>