reimplement AbstractExpirationPolicy
Rather than limit ourselves to a single timeout
length for each implementation of the timeout
logic we're going to limit the number of instances
of each class we have to the number of timeout
lengths we need. This is likely a premature
optimization but it wasn't difficult to write and
should provide a very compact/stable footprint
once the code is warmed up.

Signed-off-by: Tyler Pickett <t.pickett66@gmail.com>